Refactor server route registration
