+++ apps/server/src/routes/media.ts
@@ -1,0 +1,190 @@
+import path from "node:path";
+import { mkdir, readFile, stat, writeFile } from "node:fs/promises";
+
+import type { FastifyInstance } from "fastify";
+import type { Pool } from "pg";
+
+import type { AgentManager } from "../agents/manager.js";
+import {
+  isMediaFile,
+  isTextFile,
+  isValidMediaKey,
+  listMediaFiles,
+  loadSeenMediaKeys,
+  markSeenMediaKeys,
+  mimeType,
+  resolveMediaDir,
+  sanitizeUploadedFileName,
+  toMediaKey,
+} from "../shared/media.js";
+
+type MediaRouteDeps = {
+  pool: Pool;
+  mediaRoot: string;
+  agentManager: AgentManager;
+  publishUiEvent: (event: unknown) => void;
+};
+
+export async function registerMediaRoutes(
+  app: FastifyInstance,
+  deps: MediaRouteDeps
+): Promise<void> {
+  app.get("/api/v1/agents/:id/media", async (request, reply) => {
+    const params = request.params as { id?: string };
+    const id = params.id ?? "";
+    const agentExists = await deps.pool.query(
+      "SELECT 1 FROM agents WHERE id = $1",
+      [id]
+    );
+    if (agentExists.rows.length === 0) {
+      return reply.code(404).send({ error: "Agent not found." });
+    }
+
+    const files = await listMediaFiles(deps.pool, id);
+    const seenKeys = await loadSeenMediaKeys(
+      deps.pool,
+      id,
+      files.map(toMediaKey)
+    );
+    return {
+      files: files.map((file) => ({
+        ...file,
+        seen: seenKeys.has(toMediaKey(file)),
+      })),
+    };
+  });
+
+  app.get("/api/v1/agents/:id/media/:file", async (request, reply) => {
+    const params = request.params as { id?: string; file?: string };
+    const id = params.id ?? "";
+    const agentRow = await deps.pool.query<{
+      id: string;
+      media_dir: string | null;
+    }>("SELECT id, media_dir FROM agents WHERE id = $1", [id]);
+    if (agentRow.rows.length === 0) {
+      return reply.code(404).send({ error: "Agent not found." });
+    }
+
+    const file = params.file ?? "";
+    if (!/^[A-Za-z0-9._-]+$/.test(file)) {
+      return reply.code(400).send({ error: "Invalid media file name." });
+    }
+
+    const filePath = path.join(
+      resolveMediaDir(id, agentRow.rows[0].media_dir, deps.mediaRoot),
+      file
+    );
+    const fileStat = await stat(filePath).catch(() => null);
+    if (!fileStat || !fileStat.isFile()) {
+      return reply.code(404).send({ error: "Media file not found." });
+    }
+
+    return reply.type(mimeType(file)).send(await readFile(filePath));
+  });
+
+  app.post("/api/v1/agents/:id/media", async (request, reply) => {
+    const params = request.params as { id?: string };
+    const id = params.id ?? "";
+    const agent = await deps.agentManager.getAgent(id);
+    if (!agent) {
+      return reply.code(404).send({ error: "Agent not found." });
+    }
+
+    const data = await request.file();
+    if (!data) {
+      return reply.code(400).send({ error: "A file field is required." });
+    }
+
+    const fileName = sanitizeUploadedFileName(path.basename(data.filename));
+    if (!fileName) {
+      return reply.code(400).send({ error: "Invalid file name." });
+    }
+    if (!isMediaFile(fileName)) {
+      return reply.code(400).send({
+        error:
+          "Unsupported file type. Use images (png/jpg/gif/webp), video (mp4), documents (pdf), or text files (txt/md/json/yaml/ts/py/etc).",
+      });
+    }
+
+    const isText = isTextFile(fileName);
+    const sourceField =
+      (data.fields.source as { value?: string } | undefined)?.value ??
+      (isText ? "text" : "screenshot");
+    const validSources = ["screenshot", "stream", "simulator", "text", "user"];
+    const source = validSources.includes(sourceField)
+      ? sourceField
+      : isText
+        ? "text"
+        : "screenshot";
+    const description =
+      (data.fields.description as { value?: string } | undefined)?.value ??
+      null;
+
+    const mediaDir = resolveMediaDir(agent.id, agent.mediaDir, deps.mediaRoot);
+    await mkdir(mediaDir, { recursive: true });
+
+    const buffer = await data.toBuffer();
+    const timestamp = new Date()
+      .toISOString()
+      .replace(/[:.]/g, "-")
+      .replace("T", "-")
+      .replace("Z", "");
+    const ext = path.extname(fileName);
+    const base = path.basename(fileName, ext);
+    const timestampedFileName = `${base}-${timestamp}${ext}`;
+
+    await writeFile(path.join(mediaDir, timestampedFileName), buffer);
+
+    const result = await deps.pool.query<{ id: number; created_at: Date }>(
+      `INSERT INTO media (agent_id, file_name, source, size_bytes, description)
+       VALUES ($1, $2, $3, $4, $5)
+       RETURNING id, created_at`,
+      [id, timestampedFileName, source, buffer.length, description]
+    );
+
+    deps.publishUiEvent({ type: "media.changed", agentId: id });
+    return reply.code(201).send({
+      ok: true,
+      media: {
+        id: result.rows[0].id,
+        fileName: timestampedFileName,
+        source,
+        sizeBytes: buffer.length,
+        createdAt: result.rows[0].created_at.toISOString(),
+        url: `/api/v1/agents/${id}/media/${encodeURIComponent(timestampedFileName)}`,
+      },
+    });
+  });
+
+  app.post("/api/v1/agents/:id/media/seen", async (request, reply) => {
+    const params = request.params as { id?: string };
+    const id = params.id ?? "";
+    const agent = await deps.agentManager.getAgent(id);
+    if (!agent) {
+      return reply.code(404).send({ error: "Agent not found." });
+    }
+
+    const body = request.body as { keys?: unknown } | undefined;
+    if (
+      !Array.isArray(body?.keys) ||
+      !body.keys.every((key) => typeof key === "string")
+    ) {
+      return reply
+        .code(400)
+        .send({ error: "keys must be an array of strings." });
+    }
+
+    const keys = Array.from(
+      new Set(
+        body.keys.map((key) => key.trim()).filter((key) => isValidMediaKey(key))
+      )
+    );
+    if (keys.length === 0) {
+      return { ok: true, updated: 0 };
+    }
+
+    await markSeenMediaKeys(deps.pool, id, keys);
+    deps.publishUiEvent({ type: "media.seen", agentId: id, keys });
+    return { ok: true, updated: keys.length };
+  });
+}
